User's Guide: rewrite Pacemaker chapter for 8.4, drop Heartbeat chapter
Heartbeat (as a cluster manager, not communications layer) has been
obsolete for so long that it's safe to assume a DRBD 8.4 user will
not be using it and opt for Pacemaker instead.

Also, the relevant information for setting up a Heartbeat/Pacemaker
cluster is now in the Linux-HA User's Guide and other documentation.

Thus:

- Remove the separate chapter on Heartbeat
- Move the information worth preserving (configuring dopd) into the
  Pacemaker chapter
- Drop ancient backreferences from Pacemaker to Heartbeat