UG9: avoid crazy output of pveversion
On a switch from proxmox DM to linbit DM, we replace proxmox DM, but it
can happen that the config is kept (dpkg "rc" of the old drbdmanage).
This then confuses this stupid "pveversion" script, that still, even
though we are kicked, looks for "drbdmanage" and bullshits about the
package not cleanly installed (because in rc state).