pdf-style.yml: switch to unifont for code blocks
Changed the font being used for code blocks from the system's courier
font, to this GNU unifont project's font, which includes block drawing
glyphs (used by lsblk). This font is supposed to be one of the most
complete GNU fonts out there, so I think we should use it instead of
Courier, but the downside is it doesn't look as clean.

[rck]
Make sure the font is installed (or you use a recent checkout of
linbit-fonts).

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>