UG9: initial linstor-guide
This is a very first step to have 2 guide, the DRBD9 UG, and a separate
LINSTOR guide. As these guides are mostly shared, keep them in this
directory.

The only drawback is that we do not strictly split all the generation
steps like generating the output pngs. Even if one guide does not use an
image, it is generated (but only once). I think that is okay.