UG9,kubernetes: add a remark about metadata.initializers
This has been deprecated since 1.13, and Kubernetes >1.16 complains
about it. I have submitted a PR[1] to Stork to have it removed, but I
have no idea how long it will take to merge. Until then, let's just add
a remark here. We can remove it once the PR is merged.

[1] https://github.com/libopenstorage/stork/pull/620