UG9: remove instructions referring to drbdmanage
This may seem a little radical, but drbdmanage has been EOL since 2018,
and it doesn't really have any users anymore – so these instructions are
unlikely to be very useful for anyone.

On the flip side, it is pretty easy to stumble upon these old
instructions and wonder why they don't work (for example, while setting
up DRBD + Proxmox, as just happened to a customer).