UG9,vsan: general clarity improvements
The initial goal here was to fix the admonition labels (NOTE, TIP,
WARNING, ...). Turns out they render fine locally but not so well on our
website.

It also turns out it sometimes helps to re-read a document after some
time, since I found myself confused a few times when reading through it.
So let's make it clearer as a whole.