UG9,LUG: pdf title page settings in stylesheet
At one point we needed to adjust the placement of LINBIT's logo on the
title page within the UG9/LUG adoc headings. This is also set in the
stylesheet. At some point recently, likely after updates in asciidoc,
having this setting in both places caused conflicts. This commit removes
the logo placement from the LUG/UG9 headers and leaves this control up
to the stylesheet.