UG9: Truck based replication: Support external meta data.
Currently the script in the user's guide has meta data location
hardcoded as internal. This is now a script parameter and also
might be flex-external for example. Also the sh-ll-dev call
in writing meta data back has been replaced by sh-md-dev (which
might be different if meta data is external).

Also I have added a howto what to change in the script before
running it.