ci: render previews of the user's guides
When making changes to the User's Guides, it can sometimes be difficult
to predict what the final result will look like on the LINBIT website.
It is possible to build the asciidoc document locally, but that does not
match the style we use on the LINBIT website.

Thus, create an approximation here in our CI environment. This does most
of the post-processing that is done before publishing the User's Guides
to the website. This allows an editor to preview their changes before
they are actually published.