UG9 en: move a space at the end of previois line to avoid broken sentence.
If a sentence is split to two lines like the following (added a space at the top of line)

NVMe devices can be accessed. As the NVMe
 devices are accessed by a dedicated.

This is handled as a different msgid (maybe current po4a limitation).
   msgid "NVMe devices can be accessed. As the NVMe"
   msgid "devices are accessed by a dedicated."
However we want to have one msgid like
   msgid
   "NVMe devices can be accessed. As the NVMe "
   "devices are accessed by a dedicated."
for easy translation. Workaround for this is that a space at the top of
line is moved at the end of previous line.

NVMe devices can be accessed. As the NVMe
devices are accessed by a dedicated.

Signed-off-by: Osamu Sayama <o-sayama@sios.com>