UG9,PVE: type notify
In PVE setups we need to be careful to start the controller and
satellite before the pvedaemon. Parts of that have been committed to the
linstor-server repo, the PVE specific satellite configuration is
documented here. Actually it is not that specific to Proxmox but other
projects don't run into that problem for reasons.