dug: remove mention of nested LVM
LVM should not be used underneath DRBD *and* on top of DRBD.
Do not put an LVM physical volume underneath and on top of DRBD.
Nesting LVM will likely cause LVM tools to deadlock.
If you need to do something similar to nesting LVM, put a VM on top of DRBD, and inside that VM you can do
whatever you want.