linstor,k8s: set namespace for Helm commands
Some of our instructions refer to the linbit-sds namespace. By default, Helm
will install into the current default namespace, which almost always will be
a different namespace.

In order to sync the kustomize instructions with Helm instructions, use
a Helm command that also creates the Operator in the "linbit-sds" namespace.