UG9,reactor: rm id
A long time ago the "id" was used by reactor as a "handle" to determine
which plugins (i.e., plugin threads) to keep on a reload. Fortunately
this is gone a very long time ago, "id"s are no longer used, reactor now
compares the actual configuration and restarts plugins on reload if
their config changed (or it is a new config).